==> Figure4/08_DEGs_and_EnrichmentAnalysis/02_GSEA.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(parallel)
library(fgsea)
library(data.table)
source('/home//jupyter/BRI_Figures_Final_V2/helper_function/helper_function_IHA.r')

df<-read.csv("Deseq2_Result_Y2020-2021_Y2021-2022_D7.csv")

keep_celltypes <- c("Activated memory B cell", 
                    "CD27+ effector B cell", 
                    "CD27- effector B cell", 
                    "CD95 memory B cell",
                    "Core memory B cell",
                    "Core naive B cell", 
                    "Early memory B cell",
                    "ISG+ naive B cell",
                    "Plasma cell",
                    "Transitional B cell", 
                    "Type 2 polarized memory B cell")

df_filtered<-df%>% filter(celltype %in% keep_celltypes)

df_filtered<-df_filtered %>%
    dplyr::mutate(rank_metric = ifelse(is.na(padj), NA, -log10(pvalue) * sign(log2FoldChange)))%>%
    filter(!is.na(rank_metric))



all_combined_dfs <- list()
pathway_df <- list()
all_combined_dfs_ALL<-list()
# Loop over each dataframe in deseq2_results
for (cell_type in unique(df_filtered$celltype)) {
  print(cell_type)

  x <- df_filtered %>% filter(celltype==cell_type)
  rownames(x)<-x$gene

  ranked_genes <- setNames(x$rank_metric, rownames(x))
  
  # Load pathway file
  pathways <- gmtPathways("Hallmark_pathways.gmt")


  # Run GSEA
  set.seed(42)
  fgseaResults <- fgsea(pathways = pathways, stats = ranked_genes)
  significantPathways <- fgseaResults[fgseaResults$padj < 0.05,]
  ALLPathways <- fgseaResults
  # Skip if no significant pathways
  if (nrow(significantPathways) == 0) next
  
  df <- as.data.frame(significantPathways)
  df <- setorder(df, -NES)
  df <- as.data.frame(df)
    print(1)
  pathway_df[[cell_type]] <- df
  
  # Make a dataframe of the leading edge genes
  pathway_names <- df$pathway
  leadingEdges <- df$leadingEdge
    
  # Filter out leading edges with length less than 6
  filteredIndices <- sapply(leadingEdges, function(le) length(unlist(le)) >= 6)
  filteredPathwayNames <- pathway_names[filteredIndices]
  filteredLeadingEdges <- leadingEdges[filteredIndices]
                          
  significantPathways$celltype<-    cell_type                      
ALLPathways$celltype<-    cell_type      
  
  # Add the combinedDf to the list
  all_combined_dfs[[cell_type]] <- significantPathways
  all_combined_dfs_ALL[[cell_type]] <- ALLPathways                          
}
                        

df<-do.call(rbind,all_combined_dfs)
df$leadingEdge<-as.character(df$leadingEdge)
write.csv(df,"gsea_result.csv")

df<-do.call(rbind,all_combined_dfs_ALL)
df$leadingEdge<-as.character(df$leadingEdge)
write.csv(df,"gsea_result_all.csv")

df_filtered_cd27neg_effector_B<-df_filtered%>% filter(celltype=="CD27- effector B cell")

write.csv(df_filtered_cd27neg_effector_B,"df_filtered_cd27neg_effector_B.csv")
